+++ src/WebService/Controllers/Api/EventsController.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Mvc;
 
 using Calendar.ObjectModel.DataProviders;
 
-namespace Calendar.WebService.Controllers
+namespace Calendar.WebService.Controllers.Api
 {
     /// <summary>
     /// Provides REST endpoints for resources about events.
