--- src/WebService/Views/UserView/ViewUser.cshtml
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<!-- TODO: replace all-caps text with dynamic data. -->
-<head>
-  <meta charset="utf-8">
-  <title>DISPLAY_NAME's profile</title>
-</head>
-
-<body>
-  <h1>DISPLAY_NAME's profile</h1>
-  <div id="events">
-    <h2>DISPLAY_NAME's upcoming events</h2>
-    <ul>
-      <li><a href="../events/1.html">EVENT_TITLE</a></li>
-    </ul>
-  </div>
-  <div id="contacts">
-    <h2>DISPLAY_NAME's contacts</h2>
-    <ul>
-    </ul>
-  </div>
-</body>
-
-</html>
